requirements.txt
fastapi
uvicorn
pandas
joblib
python-dotenv
scikit-learn
tabpfn-client
numpy
